Rename api_client.py to client.py
